Refactor: changed moredetails package structure from moredetails.home.* to moredetails.* (i.e. deleted intermediate package 'home')
