+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".moredetails.home.view.MoreDetailsViewActivity"
+        <activity android:name="ayds.lisboa.songinfo.moredetails.view.MoreDetailsViewActivity"
            android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
