+++ app/src/main/java/ayds/lisboa/songinfo/home/view/HomeView.kt
@@ -16,7 +16,7 @@ import ayds.lisboa.songinfo.home.model.entities.EmptySong
 import ayds.lisboa.songinfo.home.model.entities.Song
 import ayds.lisboa.songinfo.home.model.entities.SpotifySong
 import ayds.lisboa.songinfo.home.view.HomeUiState.Companion.DEFAULT_IMAGE
-import ayds.lisboa.songinfo.moredetails.home.view.MoreDetailsViewActivity
+import ayds.lisboa.songinfo.moredetails.view.MoreDetailsViewActivity
 import ayds.lisboa.songinfo.utils.UtilsInjector
 import ayds.lisboa.songinfo.utils.navigation.NavigationUtils
 import ayds.lisboa.songinfo.utils.view.ImageLoader
