--- app/src/main/java/ayds/lisboa/songinfo/moredetails/home/model/repository/external/proxies/ProxyService.kt
@@ -1,7 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.home.model.repository.external.proxies
-
-import ayds.lisboa.songinfo.moredetails.home.model.entities.Card
-
-interface ProxyService {
-    fun getCard(artist:String): Card?
-}
