--- app/src/main/java/ayds/lisboa/songinfo/moredetails/home/model/repository/local/cards/ArtistsLocalStorage.kt
@@ -1,8 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.home.model.repository.local.cards
-
-import ayds.lisboa.songinfo.moredetails.home.model.entities.Card
-
-interface LocalStorage {
-    fun insertCards(cards: List<Card>)
-    fun getCardsByName(term: String): List<Card>
-}
