--- app/src/main/java/ayds/lisboa/songinfo/moredetails/home/view/MoreDetailsEvent.kt
@@ -1,7 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.home.view
-
-sealed class MoreDetailsEvent {
-    object SearchArtist : MoreDetailsEvent()
-    object OpenURL: MoreDetailsEvent()
-    object NextCard: MoreDetailsEvent()
-}
