--- app/src/main/java/ayds/lisboa/songinfo/moredetails/home/view/MoreDetailsState.kt
@@ -1,12 +1,0 @@
-package ayds.lisboa.songinfo.moredetails.home.view
-
-import ayds.lisboa.songinfo.moredetails.home.model.entities.Source
-
-data class MoreDetailsState (
-    val artist: String = "",
-    val bio: String = "",
-    val url: String = "",
-    val source: Source = Source.UNDEFINED,
-    val imageUrl: String = ""
-
-)
