+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/model/repository/external/proxies/ProxyService.kt
@@ -1,0 +1,7 @@
+package ayds.lisboa.songinfo.moredetails.model.repository.external.proxies
+
+import ayds.lisboa.songinfo.moredetails.model.entities.Card
+
+interface ProxyService {
+    fun getCard(artist:String): Card?
+}
