+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/model/repository/external/proxies/WikipediaProxy.kt
@@ -1,0 +1,29 @@
+package ayds.lisboa.songinfo.moredetails.model.repository.external.proxies
+
+import ayds.lisboa.songinfo.moredetails.model.entities.Card
+import ayds.lisboa.songinfo.moredetails.model.entities.CardImpl
+import ayds.lisboa.songinfo.moredetails.model.entities.Source
+import ayds.winchester2.wikipedia.ExternalRepository
+import ayds.winchester2.wikipedia.WikipediaArticle
+
+internal class WikipediaProxy(private val wikipediaService: ExternalRepository) : ProxyService {
+
+    override fun getCard(artist: String): Card? {
+
+        return try {
+            createCard(wikipediaService.getArtistDescription(artist), artist)
+        } catch (e: Exception) {
+            null
+        }
+    }
+
+    private fun createCard(serviceInfo: WikipediaArticle, artist: String) =
+        CardImpl(
+            artist,
+            serviceInfo.source,
+            serviceInfo.description,
+            false,
+            Source.WIKIPEDIA,
+            serviceInfo.sourceLogo
+        )
+}
