+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/model/repository/local/cards/ArtistsLocalStorage.kt
@@ -1,0 +1,8 @@
+package ayds.lisboa.songinfo.moredetails.model.repository.local.cards
+
+import ayds.lisboa.songinfo.moredetails.model.entities.Card
+
+interface LocalStorage {
+    fun insertCards(cards: List<Card>)
+    fun getCardsByName(term: String): List<Card>
+}
