+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/view/MoreDetailsEvent.kt
@@ -1,0 +1,7 @@
+package ayds.lisboa.songinfo.moredetails.view
+
+sealed class MoreDetailsEvent {
+    object SearchArtist : MoreDetailsEvent()
+    object OpenURL: MoreDetailsEvent()
+    object NextCard: MoreDetailsEvent()
+}
