+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/view/MoreDetailsState.kt
@@ -1,0 +1,12 @@
+package ayds.lisboa.songinfo.moredetails.view
+
+import ayds.lisboa.songinfo.moredetails.model.entities.Source
+
+data class MoreDetailsState (
+    val artist: String = "",
+    val bio: String = "",
+    val url: String = "",
+    val source: Source = Source.UNDEFINED,
+    val imageUrl: String = ""
+
+)
