+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/view/MoreDetailsViewInjector.kt
@@ -1,0 +1,15 @@
+package ayds.lisboa.songinfo.moredetails.view
+
+import ayds.lisboa.songinfo.moredetails.controller.MoreDetailsControllerInjector
+import ayds.lisboa.songinfo.moredetails.model.MoreDetailsModelInjector
+
+
+object MoreDetailsViewInjector {
+
+    val artistBioParser: ArtistBioParser = ArtistBioParserImpl()
+
+    fun init(moreDetailsView: MoreDetailsView) {
+        MoreDetailsModelInjector.initMoreDetailsModel(moreDetailsView)
+        MoreDetailsControllerInjector.onViewStarted(moreDetailsView)
+    }
+}
