+++ app/src/test/java/ayds/lisboa/songinfo/moredetails/model/repository/external/proxies/WikipediaProxyTest.kt
@@ -1,0 +1,40 @@
+package ayds.lisboa.songinfo.moredetails.model.repository.external.proxies
+
+import ayds.lisboa.songinfo.moredetails.model.entities.CardImpl
+import ayds.lisboa.songinfo.moredetails.model.entities.Source
+import ayds.winchester2.wikipedia.WikipediaArticle
+import ayds.winchester2.wikipedia.ExternalRepository as WikipediaService
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.verify
+import org.junit.Assert.*
+import org.junit.Test
+
+class WikipediaProxyTest {
+
+    private val service: WikipediaService = mockk()
+    private val proxyService: ProxyService = WikipediaProxy(service)
+
+    @Test
+    fun `given artist name should return card`() {
+        val wikipediaArticle = WikipediaArticle("infoUrl", "description", "sourceLogoUrl")
+        every { service.getArtistDescription("name") } returns wikipediaArticle
+
+        val result = proxyService.getCard("name")
+        val expectedCard = CardImpl("name", "infoUrl", "description", false, Source.WIKIPEDIA, "sourceLogoUrl")
+
+        assertEquals(expectedCard, result)
+        verify { service.getArtistDescription("name") }
+    }
+
+    @Test
+    fun `given service exception should return null`() {
+        every { service.getArtistDescription("name") } throws Exception()
+
+        val result = proxyService.getCard("name")
+
+        assertEquals(null, result)
+        verify { service.getArtistDescription("name") }
+    }
+
+}
